feat(ci): move four job names onto their gate labels and wire the façades
Four quality.yml jobs proved a registry-named property and never read the
declaration, and they could not be wired the way their siblings were, because
`contextsFor()` derives `🔍 Quality Checks / <gate.label>` and the label was not
the string the job posts. Declaring any of the four `required` produced a
branch-protection context nothing ever posts, and the pull request blocked
forever. Adding a façade first would not have fixed that — it would have armed
it.

Owner's ruling: the job name moves to the gate label. Gates do not carry vendor
names, and the vendor becomes an implementation detail of the job's steps.

  🔐 GitGuardian Secret Detection  ->  🔐 Credential Leakage
  📜 FOSSA License Check           ->  📜 Licence Check
  🔍 SonarCloud SAST               ->  🔍 Static Security Analysis
  📱 Maestro Cloud E2E Tests       ->  📱 Maestro Native E2E

`SAST` was wrong on the property as well as on the vendor: that job's verdict is
a server-side quality gate spanning reliability, maintainability, duplication
and coverage as well as security, so a duplication regression used to redden a
check whose name claimed static security analysis.

Each job then gained the byte-identical façade the other twenty-one carry, and
every pre-existing step moved onto the fallback path — the token probes with the
rest, because a probe exists to decide whether the VENDOR can run and a project
that named its own prover is not using the vendor.

The SonarCloud scan step keeps its `continue-on-error`, which is the one
pre-existing carrier and stays pinned by name. What is new is that it can only
ever be on the fallback path: a second façade assertion now proves no step
conditioned on the DECLARED path carries `continue-on-error`, so a project that
says `required` cannot inherit a step that goes green having analysed nothing.

The rename reaches every surface that names these jobs, not just the workflow:
the Rails twin workflow, the security summary strings, the console's model of
the CI jobs, the setup skill in all six agent plugins, and the wiki's job tree.
Two prose examples of "confusable required contexts" cited the SonarCloud pair;
that pair stopped being confusable when the job stopped naming the vendor, so
they now cite the GitGuardian one — an app-posted context that IS required here
sitting beside a job-posted one that is not, for the same property.

Measured blast radius, not re-derived: across the portfolio, 116 readable
rulesets and 13 unreadable, with 9 required-context entries naming any of these
four. None of them is on this repository, so no ruleset moves here. The
portfolio rollout stays deferred.

Counts: 26 QUALITY_JOB_GATES rows, from 22. Five jobs remain exempt, each with a
reason and the issue that decides it.

Work-Item: CodySwannGT/lisa#2914
