+++ all/copy-overwrite/scripts/lisa-gates.mjs
@@ -674,6 +674,15 @@ export const QUALITY_JOB_GATES = Object.freeze({
   e2e_coverage: "journey-coverage",
   state_classification: "state-classification",
   floor_collisions: "security-floor-integrity",
+  // These four could not be wired until their jobs were renamed onto their
+  // gates' labels. `contextsFor()` derives `🔍 Quality Checks / <label>`, and
+  // while each job posted a different string, declaring the gate `required`
+  // derived a context nothing ever posts and blocked the pull request forever.
+  // The façade was never the missing piece — adding it would have armed that.
+  secret_scanning: "credential-leakage",
+  license_compliance: "license-compliance",
+  maestro_e2e: "e2e-native",
+  sonarcloud: "static-security",
 });
 
 /**
@@ -712,28 +721,9 @@ export const UNGATED_QUALITY_JOBS = Object.freeze({
       "A meta-gate: it governs the governance rather than the software, alongside `gate_config_validity`, which is deliberately exempt for the same reason. Declaring it `off` would mean 'I may silence a required check without anyone objecting', which is close to self-defeating, so whether it should be declarable at all is an owner ruling and not an implementation gap.",
     owner: "#2933",
   }),
-  license_compliance: Object.freeze({
-    reason:
-      "A registry gate names this property; only the wiring is missing. Blocked on the job name and the gate label being the same string — `contextsFor()` derives the context from the label, and today the job posts a different one.",
-    owner: "#2830",
-  }),
-  maestro_e2e: Object.freeze({
-    reason: "A registry gate names this property; only the wiring is missing.",
-    owner: "#2830",
-  }),
-  secret_scanning: Object.freeze({
-    reason:
-      "A registry gate names this property; only the wiring is missing. Blocked on the job name and the gate label being the same string.",
-    owner: "#2830",
-  }),
   snyk: Object.freeze({
     reason:
       "Which property this job certifies is undecided. It is a dependency scanner, but `dependency-vulnerability` is already posted by `npm_security_scan` under a context that is required on this repository's ruleset, so two jobs would post one name.",
-    owner: "#2830",
-  }),
-  sonarcloud: Object.freeze({
-    reason:
-      "The job's verdict is a server-side quality gate spanning reliability, maintainability, duplication, coverage and security, and the conditions are held outside this repository. Mapping it onto `static-security` would let a coverage drop redden a security gate.",
     owner: "#2830",
   }),
   zap_baseline: Object.freeze({
