+++ src/core/lisa-owned-hash-ledger.ts
@@ -158,6 +158,7 @@ export const LISA_OWNED_HASH_LEDGER: Readonly<
     "0a8b14790f38e00e5abfc665d3e3d2c29e0a139a2a22c4a65e0c08547b42fb7f",
     "1bcd3f065465731df4eb02b93b73a0162d287e365ed1a396138d97d71d403b31",
     "2bafd97867aeabbb9f584e35312a7c16e2c8ad899b39152b8606fb8fff550fdb",
+    "3dc5c8c57ff052e29579bf4ba3900b2e65872c6d24177c39acd6fd830b051ad2",
     "444210c8fbfca1d3e6ae0e180fe95752e10152610ad349ffea377136f4ae8cf7",
     "48d3fa3043e144bcf52feb0e20db887d75c7ebda3b1fa708fe93f6554f6bec78",
     "5af5f112ba258c4874a65c72c8bed12165e75e4df407265eb46ee9c89c4e0e53",
@@ -289,6 +290,7 @@ export const LISA_OWNED_HASH_LEDGER: Readonly<
     "a8af7e575c086970a50cb389a159946ce1755318035ca2baecd1486f2c5b7dbc",
     "bba52f26e70e0dc399a8ed8ea9cbd3885a18f08401b28bd18871a1af4fbca375",
     "c5643655f16b3229646b59cf2ac74edc631959825231a8cf69575cc8268ce0c2",
+    "c950082c3a0e7e987a8d32455eb16103cd0516c9b907f551a93395eb2e61d253",
     "cc4270e365f53c60a5eefb58a8ac8e22be184c0821fcdcc8fe67bcfc1df5e3dd",
     "d404a7f78facf07abf5633e3aeb9ba779334bfebd2989cc4cb38d1b37d96a2ba",
     "dbca450e5a17133841ab015e71808feb028996f8297fda362dbeca4f07af2352",
@@ -395,6 +397,7 @@ export const LISA_OWNED_HASH_LEDGER: Readonly<
     "3c062d0b79961a9a93a822507e893dcdaa71d65d735557727d20464c2489d812",
     "4b383ff960d7da53e0dd46e37632ef55c5a38b004fb4aabfe91a25a9fe02d88f",
     "9b613888c4c1cc6cc2b1fb6e162a4c766012eab0ec4f8b7f2ac1bc3da6622c97",
+    "d429b5c3ad47553a73cdb2e0ea3ee81268df85d5374b06a93233d1971f7d15da",
   ]),
   "scripts/lisa-run-gates.mjs": Object.freeze([
     "080c69ab998309a0a8b4f3be74c445056da02f1403c5654bea13477fc3178604",
