+++ ui/index.html
@@ -9790,7 +9790,7 @@
                 },
               ],
               [
-                { t: "mono", v: "🔍 SonarCloud SAST" },
+                { t: "mono", v: "🔍 Static Security Analysis" },
                 { t: "text", v: "Static analysis via SonarCloud" },
                 {
                   t: "status",
@@ -10178,7 +10178,7 @@
               ],
               [
                 { t: "mono", v: "SONAR_TOKEN" },
-                { t: "text", v: "SonarCloud SAST" },
+                { t: "text", v: "Static Security Analysis" },
                 { t: "badge", v: "set", cls: "on" },
               ],
               [
